major parser test rewrite
this commit re-implements *all* of the tests for the parser. with most
of the core logic in place in the parser i thought it was a good
approach to use TDD and walk to the program and see that all of the
tests would pass. i sort of came to this conclusion after just a small
idea that i should pass another arg to the parser that would let me
debug only one test at the time, meanwhile still having all of the tests
running all of the time. this is a really nice setup as im catching all
of the errors as i do changes to the pest parser!