parser is working on real websites
this commit adds a couple of important rules to the parser:
- support for svg-elements
- support for xml-processing-instructions
- support for attributes without quotes

ive also added a cleanup rule for dangling elements, since it's common
to have these errors in production. ive also added a rule for whitespace
as a node, not sure that if this is good or not, but it's caching some
errors that i had. what is left to do now is to construct the actual
ast.