started to implment the ast
this commit starts to implment the ast, and as a first step im only
working with the simple html page and try to build each elements. i'll
continue to add the content of element with raw text, and then update
all the tests so that we can match agains a proper ast tree.