fixed derive macro bug
this commit fixed a derive macro bug that appeared when trying to use
the library as a dependency. serde didnt have the feature derive
enabled.

i also took the opportunity to expose element and elementvariant, and
removed all the patches from dependencies, to auto update dependencies
when bugfixes are published