Fix comment parsing
Change the grammar to emit a token for a comment, with a separate token
containing the text of the comment.  Modify build_dom and
build_node_element to handle these new tokens.

In JSON format, text and comments are indistinguishable.  But the code
has been tested to prove that it does emit different Node enum variants.