Fix comment parsing (#12)
* Fix comment parsing

Change the grammar to emit a token for a comment, with a separate token
containing the text of the comment.  Modify build_dom and
build_node_element to handle these new tokens.

In JSON format, text and comments are indistinguishable.  But the code
has been tested to prove that it does emit different Node enum variants.

* Update test JSON snapshots to account for comments

Now that the parser properly generates Node elements for comments, the
tests need to be updated to account for that.

* Fix tree_type selection logic; validation rewrite

This commit makes some significant changes to how the tree_type is
selected.  Previously, this was selected based entirely on the number of
children in the root of the dom.  Now that we can properly parse
comments (and create the proper Node::Comment() children), that isn't a
valid way to determine document type.  The new logic is extensively
documented in the comments of the file, but it boils down to "assume
Empty to start; if we see DOCTYPE then it's a document; if we see
fragments, then it is a fragment" with some fixups during the validation
stage.

The validation is also re-written, as it also depended on the idea of
"counting children" to make decisions.  The logic is documented in the
comments, but is basically broken down into three different sets of
rules depending on the tree_type.  Empty documents can have comments,
but nothing else; a Document can have comments and a single HTML tag at
the top level, and Fragments can't have HEAD or BODY (but can be
upgraded to a Document if all it has is an HTML tag).

Also fixed a test case to account for new comment parsing capabilities.