Move binary target to be an example (#11)
* Move binary target to be an example

Move src/main.rs from being a binary target for the crate to being an
example.  Correspondingly, move structopt from a dependency to a
dev-dependency.  This reduces the number of dependencies to just build
the library from 44 crates to 27 crates.

* Fix tests/bin.rs to call new example binary

The test in tests/bin.rs used to call the crate binary.  Since that has
been moved to an example, the test needed to be updated.